[cli] modularize the terraform and lambda handlers
* Per @ryandeivert's comment, I moved the Terraform clean and other functionality into the Terraform CLI module.
* As a result, other dependencies needed to be modularized to prevent circular dependencies.
* The Lambda functionality had to also be broken out as a result.
